+++ src/frontend/staff-portal/src/app/components/jj-workbench/jj-dispute/jj-dispute.component.html
@@ -1,5 +1,5 @@
-<div class="contact-info">
-  <a (click)="onBack()" style="color: #48afe5">
+<div [ngClass]="{'contact-info' : type !== 'ticket'}">
+  <a *ngIf="type !== 'ticket'" (click)="onBack()" style="color: #48afe5">
     <mat-icon inline="true">arrow_back</mat-icon> Back<br /><br />
   </a>
   <div *ngIf="retrieving === true">
@@ -148,59 +148,45 @@
       <textarea matInput [(ngModel)]="lastUpdatedJJDispute.remarks" maxlength="1500" rows="6"
         style="background: #fffee6;"></textarea>
     </section>
-    <section *ngIf="type === 'ticket'">
-      <ng-container subHeader>
-        <span>
-          <h3>JJ Remarks</h3>
-        </span>
-      </ng-container>
-      <div>{{ lastUpdatedJJDispute.remarks }}</div>
-    </section>
-    <br />
-    <div fxLayout="row">
-      <div fxFlex="40">
-        <div fxFlex>
-          <h3>Signed By</h3>
-          <div>{{ lastUpdatedJJDispute.jjAssignedTo }}</div>
-        </div>
-        <div fxFlex></div>
-        <div fxFlex>
-          <h3>Date Signed</h3>
-          <div>{{ lastUpdatedJJDispute.jjDecisionDate | date: "MMM dd, yyyy" }}</div>
-        </div>
-      </div>
-    </div>
-    <section *ngIf="type === 'ticket'" style="padding-top: 2rem;">
-      <ng-container subHeader>
-        <span>
-          <h3>Add Remarks<br />
-            <span class="small-note">
-              <i><span style="color:red;">*</span>
-                Remarks will be sent to the JJ if you need toreturn this dispute decision</i>
-            </span>
-          </h3>
-        </span>
-      </ng-container>
-      <textarea matInput maxlength="1500" rows="6" style="background: #fffee6;"></textarea>
-    </section>
-    <br />
-    <div fxLayout="row">
-      <div fxFlex="40" fxLayout="row">
-        <div style="margin-right: 1rem;">
-          <mat-form-field class="select-box-only">
-            <mat-select [(ngModel)]="selectedJJ" placeholder="choose a JJ">
-              <mat-option [value]=""></mat-option>
-              <mat-option *ngFor="let jj of jjList" [value]="jj.idir"><i>JJ {{ jj.name }}</i></mat-option>
-            </mat-select>
-          </mat-form-field>
-        </div>
-        <div fxFlex>
-          <button fxFlex *ngIf="type === 'ticket'" class="large" mat-flat-button type="button" color="primary"
-            (click)="returnToJJ()" [disabled]="!selectedJJ">
-            Return to JJ
-          </button>
+    <ng-container *ngIf="type === 'ticket'">
+      <section>
+        <ng-container subHeader>
+          <span>
+            <h3>JJ Remarks</h3>
+          </span>
+        </ng-container>
+        <div>{{ lastUpdatedJJDispute.remarks }}</div>
+      </section>
+      <br />
+      <div fxLayout="row">
+        <div fxFlex="40">
+          <div fxFlex>
+            <h3>Signed By</h3>
+            <div>{{ lastUpdatedJJDispute.jjAssignedTo }}</div>
+          </div>
+          <div fxFlex></div>
+          <div fxFlex>
+            <h3>Date Signed</h3>
+            <div>{{ lastUpdatedJJDispute.jjDecisionDate | date: "MMM dd, yyyy" }}</div>
+          </div>
         </div>
       </div>
+      <section style="padding-top: 2rem;">
+        <ng-container subHeader>
+          <span>
+            <h3>Add Remarks<br />
+              <span class="small-note">
+                <i><span style="color:red;">*</span>
+                  Remarks will be sent to the JJ if you need toreturn this dispute decision</i>
+              </span>
+            </h3>
+          </span>
+        </ng-container>
+        <textarea matInput maxlength="1500" rows="6" style="background: #fffee6;"></textarea>
+      </section>
+    </ng-container>
+    <br />
+    <div *ngIf="type !== 'ticket'" fxLayout="row">
       <div fxFlex>
       </div>
       <button fxFlex="30" *ngIf="type !== 'ticket'" class="large" mat-flat-button type="button" color="primary"
